+++ src/js/select2/data/tags.js
@@ -111,8 +111,6 @@ define([
   };
 
   Tags.prototype._removeOldTags = function (_) {
-    var tag = this._lastTag;
-
     var $options = this.$element.find('option[data-select2-tag]');
 
     $options.each(function () {