+++ src/js/select2/i18n/de.js
@@ -18,12 +18,10 @@ define(function () {
       return 'Lade mehr Ergebnisse…';
     },
     maximumSelected: function (args) {
-      var message = 'Sie können nur ' + args.maximum + ' Eintr';
+      var message = 'Sie können nur ' + args.maximum + ' Element';
 
-      if (args.maximum === 1) {
-        message += 'ag';
-      } else {
-        message += 'äge';
+      if (args.maximum != 1) {
+        message += 'e';
       }
 
       message += ' auswählen';
@@ -37,7 +35,7 @@ define(function () {
       return 'Suche…';
     },
     removeAllItems: function () {
-      return 'Entferne alle Gegenstände';
+      return 'Entferne alle Elemente';
     }
   };
 });