+++ tests/data/array-tests.js
@@ -71,6 +71,9 @@ test('current gets default for single', function (assert) {
 
   var data = new ArrayData($select, arrayOptions);
 
+  var container = new MockContainer();
+  data.bind(container, $('<div></div>'));
+
   data.current(function (val) {
     assert.equal(
       val.length,
@@ -93,6 +96,9 @@ test('current gets default for multiple', function (assert) {
 
   var data = new ArrayData($select, arrayOptions);
 
+  var container = new MockContainer();
+  data.bind(container, $('<div></div>'));
+
   data.current(function (val) {
     assert.equal(
       val.length,
@@ -106,6 +112,9 @@ test('current works with existing selections', function (assert) {
   var $select = $('#qunit-fixture .multiple');
 
   var data = new ArrayData($select, arrayOptions);
+
+  var container = new MockContainer();
+  data.bind(container, $('<div></div>'));
 
   $select.val(['One']);
 
@@ -136,6 +145,9 @@ test('current works with selected data', function (assert) {
   var $select = $('#qunit-fixture .single-empty');
 
   var data = new ArrayData($select, arrayOptions);
+
+  var container = new MockContainer();
+  data.bind(container, $('<div></div>'));
 
   data.select({
     id: '2',
@@ -170,6 +182,9 @@ test('select works for single', function (assert) {
 
   var data = new ArrayData($select, arrayOptions);
 
+  var container = new MockContainer();
+  data.bind(container, $('<div></div>'));
+
   assert.equal(
     $select.val(),
     'default',
@@ -193,6 +208,9 @@ test('multiple sets the value', function (assert) {
 
   var data = new ArrayData($select, arrayOptions);
 
+  var container = new MockContainer();
+  data.bind(container, $('<div></div>'));
+
   assert.ok(
     $select.val() == null || $select.val().length == 0,
     'nothing should be selected'
@@ -211,6 +229,9 @@ test('multiple adds to the old value', function (assert) {
 
   var data = new ArrayData($select, arrayOptions);
 
+  var container = new MockContainer();
+  data.bind(container, $('<div></div>'));
+
   $select.val(['One']);
 
   assert.deepEqual($select.val(), ['One']);
@@ -228,6 +249,9 @@ test('option tags are automatically generated', function (assert) {
 
   var data = new ArrayData($select, arrayOptions);
 
+  var container = new MockContainer();
+  data.bind(container, $('<div></div>'));
+
   assert.equal(
     $select.find('option').length,
     4,
@@ -235,10 +259,31 @@ test('option tags are automatically generated', function (assert) {
   );
 });
 
+test('automatically generated option tags have a result id', function (assert) {
+  var $select = $('#qunit-fixture .single-empty');
+
+  var data = new ArrayData($select, arrayOptions);
+
+  var container = new MockContainer();
+  data.bind(container, $('<div></div>'));
+
+  data.select({
+    id: 'default'
+  });
+
+  assert.ok(
+    Utils.GetData($select.find(':selected')[0], 'data')._resultId,
+    '<option> default should have a result ID assigned'
+  );
+});
+
 test('option tags can receive new data', function(assert) {
   var $select = $('#qunit-fixture .single');
 
   var data = new ArrayData($select, extraOptions);
+
+  var container = new MockContainer();
+  data.bind(container, $('<div></div>'));
 
   assert.equal(
     $select.find('option').length,
@@ -270,6 +315,9 @@ test('optgroup tags can also be generated', function (assert) {
 
   var data = new ArrayData($select, nestedOptions);
 
+  var container = new MockContainer();
+  data.bind(container, $('<div></div>'));
+
   assert.equal(
     $select.find('option').length,
     1,
@@ -287,6 +335,9 @@ test('optgroup tags have the right properties', function (assert) {
   var $select = $('#qunit-fixture .single-empty');
 
   var data = new ArrayData($select, nestedOptions);
+
+  var container = new MockContainer();
+  data.bind(container, $('<div></div>'));
 
   var $group = $select.children('optgroup');
 
@@ -327,6 +378,9 @@ test('existing selections are respected on initialization', function (assert) {
     assert.equal($select.val(), 'Second');
 
     var data = new ArrayData($select, options);
+
+    var container = new MockContainer();
+    data.bind(container, $('<div></div>'));
 
     assert.equal($select.val(), 'Second');
 });